+++ web/src/lib/elements/StarRating.svelte
